refactor(#2456): split data_fitting.py into models/solvers/coordinator
- Extract BodySegmentParams, KinematicState, FitResult, SensitivityResult,
  ParameterEstimationReport to _data_fitting_models.py (163 LOC)
- Extract InverseKinematicsSolver, ParameterEstimator to
  _data_fitting_solvers.py (505 LOC)
- Slim data_fitting.py coordinator to 496 LOC (was 1096 LOC)
- All public names re-exported for backward compatibility
- TDD contract tests added (15 tests, all GREEN)
